Update delete offence history data logic
Now that we have a third option for the previous convictions question,
we update the utility that removes orphaned data.

+++ server/utils/applications/deleteOrphanedData.ts
@@ -1,5 +1,6 @@
 import { AnyValue } from '@approved-premises/api'
 import { lastKnownKeys, previousKeys } from '../../form-pages/apply/about-the-person/address-history/previousAddress'
+import { PreviousConvictionsAnswers } from '../../form-pages/apply/offence-and-licence-information/offending-history/anyPreviousConvictions'
 
 export default function deleteOrphanedFollowOnAnswers(applicationData: AnyValue): AnyValue {
   const deleteOrphanedFundingInformation = () => {
@@ -68,7 +69,18 @@ export default function deleteOrphanedFollowOnAnswers(applicationData: AnyValue)
       taskName: 'offending-history',
       pageName: 'any-previous-convictions',
       questionKey: 'hasAnyPreviousConvictions',
-      answerToCheck: 'no',
+      answerToCheck: PreviousConvictionsAnswers.No,
+    })
+  ) {
+    deleteOrphanedOffendingHistoryInformation()
+  }
+
+  if (
+    hasOrphanedInformation({
+      taskName: 'offending-history',
+      pageName: 'any-previous-convictions',
+      questionKey: 'hasAnyPreviousConvictions',
+      answerToCheck: PreviousConvictionsAnswers.YesNoRelevantRisk,
     })
   ) {
     deleteOrphanedOffendingHistoryInformation()
